refactor(parser): split parser module internals
