+++ crates/shuck-parser/src/parser/mod.rs
@@ -7,9 +7,15 @@
 mod arithmetic;
 mod commands;
 mod heredocs;
+mod keywords;
 mod lexer;
+mod parser_state;
+mod profile;
 mod redirects;
+mod result;
 mod words;
+mod zsh_options;
+mod zsh_prescan;
 
 use std::{
     borrow::Cow,
@@ -21,8 +27,18 @@ pub use lexer::{
     HeredocRead, LexedToken, LexedWord, LexedWordSegment, LexedWordSegmentKind, Lexer,
     LexerErrorKind,
 };
+pub use profile::{ShellDialect, ShellProfile};
+pub use result::{ParseDiagnostic, ParseResult, ParseStatus, SyntaxFacts, ZshCaseGroupPart};
+pub use zsh_options::{OptionValue, ZshEmulationMode, ZshOptionState};
+
+use keywords::*;
 use memchr::{memchr, memchr2, memchr3};
+pub use parser_state::Parser;
+#[cfg(feature = "benchmarking")]
+pub use parser_state::ParserBenchmarkCounters;
+use parser_state::*;
 use smallvec::SmallVec;
+use zsh_prescan::ZshOptionTimeline;
 
 use shuck_ast::{
     AlwaysCommand, AnonymousFunctionCommand, AnonymousFunctionSurface, ArithmeticCommand,
@@ -41,13 +57,12 @@ use shuck_ast::{
     LiteralText, Name, ParameterExpansion, ParameterExpansionSyntax, ParameterOp, Pattern,
     PatternGroupKind, PatternPart, PatternPartNode, Position, PrefixMatchKind, Redirect,
     RedirectKind, RedirectTarget, RepeatCommand, RepeatSyntax, ReturnCommand as AstReturnCommand,
-    SelectCommand, SimpleCommand as AstSimpleCommand, SourceText, Span, StaticCommandWrapperTarget,
-    Stmt, StmtSeq, StmtTerminator, Subscript, SubscriptInterpretation, SubscriptKind,
-    SubscriptSelector, TextSize, TimeCommand, TokenKind, UntilCommand, VarRef, WhileCommand, Word,
-    WordPart, WordPartNode, ZshDefaultingOp, ZshExpansionOperation, ZshExpansionTarget,
-    ZshGlobQualifier, ZshGlobQualifierGroup, ZshGlobQualifierKind, ZshGlobSegment,
-    ZshInlineGlobControl, ZshModifier, ZshParameterExpansion, ZshPatternOp, ZshQualifiedGlob,
-    ZshReplacementOp, ZshTrimOp, static_command_wrapper_target_index,
+    SelectCommand, SimpleCommand as AstSimpleCommand, SourceText, Span, Stmt, StmtSeq,
+    StmtTerminator, Subscript, SubscriptInterpretation, SubscriptKind, SubscriptSelector, TextSize,
+    TimeCommand, TokenKind, UntilCommand, VarRef, WhileCommand, Word, WordPart, WordPartNode,
+    ZshDefaultingOp, ZshExpansionOperation, ZshExpansionTarget, ZshGlobQualifier,
+    ZshGlobQualifierGroup, ZshGlobQualifierKind, ZshGlobSegment, ZshInlineGlobControl, ZshModifier,
+    ZshParameterExpansion, ZshPatternOp, ZshQualifiedGlob, ZshReplacementOp, ZshTrimOp,
 };
 
 use crate::error::{Error, Result};
@@ -70,122 +85,6 @@ const HARD_MAX_AST_DEPTH: usize = 100;
 
 /// Default maximum parser operations (matches ExecutionLimits default)
 const DEFAULT_MAX_PARSER_OPERATIONS: usize = 100_000;
-
-/// Overall outcome of a parse attempt.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ParseStatus {
-    /// The parse completed without recovery diagnostics.
-    Clean,
-    /// The parse completed, but required recovery diagnostics.
-    Recovered,
-    /// The parse failed with a terminal error.
-    Fatal,
-}
-
-/// One branch separator recognized inside a zsh `case` pattern group.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ZshCaseGroupPart {
-    /// Index of the owning pattern part within the parsed pattern.
-    pub pattern_part_index: usize,
-    /// Source span covering the separator syntax.
-    pub span: Span,
-}
-
-/// Additional parser-owned facts that are useful to downstream consumers.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct SyntaxFacts {
-    /// Spans of zsh brace-style `if` bodies.
-    pub zsh_brace_if_spans: Vec<Span>,
-    /// Spans of zsh `always` clauses.
-    pub zsh_always_spans: Vec<Span>,
-    /// Pattern-group separators collected from zsh `case` items.
-    pub zsh_case_group_parts: Vec<ZshCaseGroupPart>,
-}
-
-/// The result of parsing a script, including any recovery diagnostics and
-/// syntax facts collected along the way.
-#[derive(Debug, Clone)]
-pub struct ParseResult {
-    /// Parsed syntax tree for the file.
-    pub file: File,
-    /// Recovery diagnostics emitted while producing the AST.
-    pub diagnostics: Vec<ParseDiagnostic>,
-    /// High-level parse status.
-    pub status: ParseStatus,
-    /// Terminal parse error, when recovery could not continue.
-    pub terminal_error: Option<Error>,
-    /// Additional syntax facts collected during parsing.
-    pub syntax_facts: SyntaxFacts,
-}
-
-impl ParseResult {
-    /// Returns `true` when the parse completed without recovery diagnostics.
-    pub fn is_ok(&self) -> bool {
-        self.status == ParseStatus::Clean
-    }
-
-    /// Returns `true` when the parse produced recovery diagnostics or a terminal error.
-    pub fn is_err(&self) -> bool {
-        !self.is_ok()
-    }
-
-    /// Convert this result into a strict parse error.
-    ///
-    /// If recovery diagnostics exist but no terminal error was recorded, the first recovery
-    /// diagnostic is converted into an [`Error`].
-    pub fn strict_error(&self) -> Error {
-        self.terminal_error.clone().unwrap_or_else(|| {
-            let Some(diagnostic) = self.diagnostics.first() else {
-                panic!("non-clean parse result should include a diagnostic or terminal error");
-            };
-            Error::parse_at(
-                diagnostic.message.clone(),
-                diagnostic.span.start.line,
-                diagnostic.span.start.column,
-            )
-        })
-    }
-
-    /// Return the parse result when it is clean, otherwise panic with the strict error.
-    pub fn unwrap(self) -> Self {
-        if self.is_ok() {
-            self
-        } else {
-            panic!(
-                "called `ParseResult::unwrap()` on a non-clean parse: {}",
-                self.strict_error()
-            )
-        }
-    }
-
-    /// Return the parse result when it is clean, otherwise panic with `message`.
-    pub fn expect(self, message: &str) -> Self {
-        if self.is_ok() {
-            self
-        } else {
-            panic!("{message}: {}", self.strict_error())
-        }
-    }
-
-    /// Return the strict parse error when the result is not clean, otherwise panic.
-    pub fn unwrap_err(self) -> Error {
-        if self.is_err() {
-            self.strict_error()
-        } else {
-            panic!("called `ParseResult::unwrap_err()` on a clean parse")
-        }
-    }
-
-    /// Return the strict parse error when the result is not clean, otherwise panic with
-    /// `message`.
-    pub fn expect_err(self, message: &str) -> Error {
-        if self.is_err() {
-            self.strict_error()
-        } else {
-            panic!("{message}")
-        }
-    }
-}
 
 /// Returns whether `text` parses as a nontrivial arithmetic expression.
 pub fn text_looks_like_nontrivial_arithmetic_expression(text: &str) -> bool {
@@ -272,1538 +171,37 @@ fn arithmetic_expr_is_self_contained(expr: &ArithmeticExprNode) -> bool {
     }
 }
 
-#[cfg(feature = "benchmarking")]
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-#[doc(hidden)]
-pub struct ParserBenchmarkCounters {
-    /// Number of lexer current-position lookups performed while parsing.
-    pub lexer_current_position_calls: u64,
-    /// Number of parser calls that updated the current spanned token.
-    pub parser_set_current_spanned_calls: u64,
-    /// Number of raw token-advance operations performed by the parser.
-    pub parser_advance_raw_calls: u64,
-}
-
-#[derive(Debug, Clone)]
-struct SimpleCommand {
-    name: Word,
-    args: SmallVec<[Word; 2]>,
-    redirects: SmallVec<[Redirect; 1]>,
-    assignments: SmallVec<[Assignment; 1]>,
-    span: Span,
-}
-
-#[derive(Debug, Clone)]
-struct BreakCommand {
-    depth: Option<Word>,
-    extra_args: SmallVec<[Word; 2]>,
-    redirects: SmallVec<[Redirect; 1]>,
-    assignments: SmallVec<[Assignment; 1]>,
-    span: Span,
-}
-
-#[derive(Debug, Clone)]
-struct ContinueCommand {
-    depth: Option<Word>,
-    extra_args: SmallVec<[Word; 2]>,
-    redirects: SmallVec<[Redirect; 1]>,
-    assignments: SmallVec<[Assignment; 1]>,
-    span: Span,
-}
-
-#[derive(Debug, Clone)]
-struct ReturnCommand {
-    code: Option<Word>,
-    extra_args: SmallVec<[Word; 2]>,
-    redirects: SmallVec<[Redirect; 1]>,
-    assignments: SmallVec<[Assignment; 1]>,
-    span: Span,
-}
-
-#[derive(Debug, Clone)]
-struct ExitCommand {
-    code: Option<Word>,
-    extra_args: SmallVec<[Word; 2]>,
-    redirects: SmallVec<[Redirect; 1]>,
-    assignments: SmallVec<[Assignment; 1]>,
-    span: Span,
-}
-
-#[derive(Debug, Clone)]
-enum BuiltinCommand {
-    Break(BreakCommand),
-    Continue(ContinueCommand),
-    Return(ReturnCommand),
-    Exit(ExitCommand),
-}
-
-#[derive(Debug, Clone)]
-struct DeclClause {
-    variant: Name,
-    variant_span: Span,
-    operands: SmallVec<[DeclOperand; 2]>,
-    redirects: SmallVec<[Redirect; 1]>,
-    assignments: SmallVec<[Assignment; 1]>,
-    span: Span,
-}
-
-#[derive(Debug, Clone)]
-enum Command {
-    Simple(SimpleCommand),
-    Builtin(BuiltinCommand),
-    Decl(Box<DeclClause>),
-    Compound(Box<CompoundCommand>, SmallVec<[Redirect; 1]>),
-    Function(FunctionDef),
-    AnonymousFunction(AnonymousFunctionCommand, SmallVec<[Redirect; 1]>),
-}
-
-/// Supported shell dialects for parsing.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum ShellDialect {
-    /// POSIX-style parsing used for `sh`, `dash`, and generic portable shell input.
-    Posix,
-    /// mksh-specific parsing.
-    Mksh,
-    /// Bash parsing.
-    #[default]
-    Bash,
-    /// zsh parsing.
-    Zsh,
-}
-
-/// Tri-state option value used when modeling zsh option state.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum OptionValue {
-    /// The option is enabled.
-    On,
-    /// The option is disabled.
-    Off,
-    /// The option value is unknown or differs across merged states.
-    #[default]
-    Unknown,
-}
-
-impl OptionValue {
-    /// Returns `true` when the option is known to be enabled.
-    pub const fn is_definitely_on(self) -> bool {
-        matches!(self, Self::On)
-    }
-
-    /// Returns `true` when the option is known to be disabled.
-    pub const fn is_definitely_off(self) -> bool {
-        matches!(self, Self::Off)
-    }
-
-    /// Merge two option values, preserving certainty only when they agree.
-    pub const fn merge(self, other: Self) -> Self {
-        match (self, other) {
-            (Self::On, Self::On) => Self::On,
-            (Self::Off, Self::Off) => Self::Off,
-            _ => Self::Unknown,
-        }
-    }
-}
-
-/// Target emulation mode for zsh's `emulate` behavior.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum ZshEmulationMode {
-    /// Native zsh behavior.
-    Zsh,
-    /// `sh` compatibility mode.
-    Sh,
-    /// `ksh` compatibility mode.
-    Ksh,
-    /// `csh` compatibility mode.
-    Csh,
-}
-
-/// Snapshot of zsh option state used by the parser and lexer.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ZshOptionState {
-    /// State of the `sh_word_split` option.
-    pub sh_word_split: OptionValue,
-    /// State of the `glob_subst` option.
-    pub glob_subst: OptionValue,
-    /// State of the `rc_expand_param` option.
-    pub rc_expand_param: OptionValue,
-    /// State of the `glob` option.
-    pub glob: OptionValue,
-    /// State of the `nomatch` option.
-    pub nomatch: OptionValue,
-    /// State of the `null_glob` option.
-    pub null_glob: OptionValue,
-    /// State of the `csh_null_glob` option.
-    pub csh_null_glob: OptionValue,
-    /// State of the `extended_glob` option.
-    pub extended_glob: OptionValue,
-    /// State of the `ksh_glob` option.
-    pub ksh_glob: OptionValue,
-    /// State of the `sh_glob` option.
-    pub sh_glob: OptionValue,
-    /// State of the `bare_glob_qual` option.
-    pub bare_glob_qual: OptionValue,
-    /// State of the `glob_dots` option.
-    pub glob_dots: OptionValue,
-    /// State of the `equals` option.
-    pub equals: OptionValue,
-    /// State of the `magic_equal_subst` option.
-    pub magic_equal_subst: OptionValue,
-    /// State of the `sh_file_expansion` option.
-    pub sh_file_expansion: OptionValue,
-    /// State of the `glob_assign` option.
-    pub glob_assign: OptionValue,
-    /// State of the `ignore_braces` option.
-    pub ignore_braces: OptionValue,
-    /// State of the `ignore_close_braces` option.
-    pub ignore_close_braces: OptionValue,
-    /// State of the `brace_ccl` option.
-    pub brace_ccl: OptionValue,
-    /// State of the `ksh_arrays` option.
-    pub ksh_arrays: OptionValue,
-    /// State of the `ksh_zero_subscript` option.
-    pub ksh_zero_subscript: OptionValue,
-    /// State of the `short_loops` option.
-    pub short_loops: OptionValue,
-    /// State of the `short_repeat` option.
-    pub short_repeat: OptionValue,
-    /// State of the `rc_quotes` option.
-    pub rc_quotes: OptionValue,
-    /// State of the `interactive_comments` option.
-    pub interactive_comments: OptionValue,
-    /// State of the `c_bases` option.
-    pub c_bases: OptionValue,
-    /// State of the `octal_zeroes` option.
-    pub octal_zeroes: OptionValue,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-enum ZshOptionField {
-    ShWordSplit,
-    GlobSubst,
-    RcExpandParam,
-    Glob,
-    Nomatch,
-    NullGlob,
-    CshNullGlob,
-    ExtendedGlob,
-    KshGlob,
-    ShGlob,
-    BareGlobQual,
-    GlobDots,
-    Equals,
-    MagicEqualSubst,
-    ShFileExpansion,
-    GlobAssign,
-    IgnoreBraces,
-    IgnoreCloseBraces,
-    BraceCcl,
-    KshArrays,
-    KshZeroSubscript,
-    ShortLoops,
-    ShortRepeat,
-    RcQuotes,
-    InteractiveComments,
-    CBases,
-    OctalZeroes,
-}
-
-impl ZshOptionState {
-    /// Default zsh option state used for native zsh parsing.
-    pub const fn zsh_default() -> Self {
-        Self {
-            sh_word_split: OptionValue::Off,
-            glob_subst: OptionValue::Off,
-            rc_expand_param: OptionValue::Off,
-            glob: OptionValue::On,
-            nomatch: OptionValue::On,
-            null_glob: OptionValue::Off,
-            csh_null_glob: OptionValue::Off,
-            extended_glob: OptionValue::Off,
-            ksh_glob: OptionValue::Off,
-            sh_glob: OptionValue::Off,
-            bare_glob_qual: OptionValue::On,
-            glob_dots: OptionValue::Off,
-            equals: OptionValue::On,
-            magic_equal_subst: OptionValue::Off,
-            sh_file_expansion: OptionValue::Off,
-            glob_assign: OptionValue::Off,
-            ignore_braces: OptionValue::Off,
-            ignore_close_braces: OptionValue::Off,
-            brace_ccl: OptionValue::Off,
-            ksh_arrays: OptionValue::Off,
-            ksh_zero_subscript: OptionValue::Off,
-            short_loops: OptionValue::On,
-            short_repeat: OptionValue::On,
-            rc_quotes: OptionValue::Off,
-            interactive_comments: OptionValue::On,
-            c_bases: OptionValue::Off,
-            octal_zeroes: OptionValue::Off,
-        }
-    }
-
-    /// Option state implied by `emulate <mode>`.
-    pub fn for_emulate(mode: ZshEmulationMode) -> Self {
-        let mut state = Self::zsh_default();
-        match mode {
-            ZshEmulationMode::Zsh => {}
-            ZshEmulationMode::Sh => {
-                state.sh_word_split = OptionValue::On;
-                state.glob_subst = OptionValue::On;
-                state.sh_glob = OptionValue::On;
-                state.sh_file_expansion = OptionValue::On;
-                state.bare_glob_qual = OptionValue::Off;
-                state.ksh_arrays = OptionValue::Off;
-            }
-            ZshEmulationMode::Ksh => {
-                state.sh_word_split = OptionValue::On;
-                state.glob_subst = OptionValue::On;
-                state.ksh_glob = OptionValue::On;
-                state.ksh_arrays = OptionValue::On;
-                state.sh_glob = OptionValue::On;
-                state.bare_glob_qual = OptionValue::Off;
-            }
-            ZshEmulationMode::Csh => {
-                state.csh_null_glob = OptionValue::On;
-                state.sh_word_split = OptionValue::Off;
-                state.glob_subst = OptionValue::Off;
-            }
-        }
-        state
-    }
-
-    /// Apply a zsh `setopt`-style option name.
-    ///
-    /// Returns `true` when the option name was recognized.
-    pub fn apply_setopt(&mut self, name: &str) -> bool {
-        self.apply_named_option(name, true)
-    }
-
-    /// Apply a zsh `unsetopt`-style option name.
-    ///
-    /// Returns `true` when the option name was recognized.
-    pub fn apply_unsetopt(&mut self, name: &str) -> bool {
-        self.apply_named_option(name, false)
-    }
-
-    fn set_field(&mut self, field: ZshOptionField, value: OptionValue) {
-        match field {
-            ZshOptionField::ShWordSplit => self.sh_word_split = value,
-            ZshOptionField::GlobSubst => self.glob_subst = value,
-            ZshOptionField::RcExpandParam => self.rc_expand_param = value,
-            ZshOptionField::Glob => self.glob = value,
-            ZshOptionField::Nomatch => self.nomatch = value,
-            ZshOptionField::NullGlob => self.null_glob = value,
-            ZshOptionField::CshNullGlob => self.csh_null_glob = value,
-            ZshOptionField::ExtendedGlob => self.extended_glob = value,
-            ZshOptionField::KshGlob => self.ksh_glob = value,
-            ZshOptionField::ShGlob => self.sh_glob = value,
-            ZshOptionField::BareGlobQual => self.bare_glob_qual = value,
-            ZshOptionField::GlobDots => self.glob_dots = value,
-            ZshOptionField::Equals => self.equals = value,
-            ZshOptionField::MagicEqualSubst => self.magic_equal_subst = value,
-            ZshOptionField::ShFileExpansion => self.sh_file_expansion = value,
-            ZshOptionField::GlobAssign => self.glob_assign = value,
-            ZshOptionField::IgnoreBraces => self.ignore_braces = value,
-            ZshOptionField::IgnoreCloseBraces => self.ignore_close_braces = value,
-            ZshOptionField::BraceCcl => self.brace_ccl = value,
-            ZshOptionField::KshArrays => self.ksh_arrays = value,
-            ZshOptionField::KshZeroSubscript => self.ksh_zero_subscript = value,
-            ZshOptionField::ShortLoops => self.short_loops = value,
-            ZshOptionField::ShortRepeat => self.short_repeat = value,
-            ZshOptionField::RcQuotes => self.rc_quotes = value,
-            ZshOptionField::InteractiveComments => self.interactive_comments = value,
-            ZshOptionField::CBases => self.c_bases = value,
-            ZshOptionField::OctalZeroes => self.octal_zeroes = value,
-        }
-    }
-
-    fn field(&self, field: ZshOptionField) -> OptionValue {
-        match field {
-            ZshOptionField::ShWordSplit => self.sh_word_split,
-            ZshOptionField::GlobSubst => self.glob_subst,
-            ZshOptionField::RcExpandParam => self.rc_expand_param,
-            ZshOptionField::Glob => self.glob,
-            ZshOptionField::Nomatch => self.nomatch,
-            ZshOptionField::NullGlob => self.null_glob,
-            ZshOptionField::CshNullGlob => self.csh_null_glob,
-            ZshOptionField::ExtendedGlob => self.extended_glob,
-            ZshOptionField::KshGlob => self.ksh_glob,
-            ZshOptionField::ShGlob => self.sh_glob,
-            ZshOptionField::BareGlobQual => self.bare_glob_qual,
-            ZshOptionField::GlobDots => self.glob_dots,
-            ZshOptionField::Equals => self.equals,
-            ZshOptionField::MagicEqualSubst => self.magic_equal_subst,
-            ZshOptionField::ShFileExpansion => self.sh_file_expansion,
-            ZshOptionField::GlobAssign => self.glob_assign,
-            ZshOptionField::IgnoreBraces => self.ignore_braces,
-            ZshOptionField::IgnoreCloseBraces => self.ignore_close_braces,
-            ZshOptionField::BraceCcl => self.brace_ccl,
-            ZshOptionField::KshArrays => self.ksh_arrays,
-            ZshOptionField::KshZeroSubscript => self.ksh_zero_subscript,
-            ZshOptionField::ShortLoops => self.short_loops,
-            ZshOptionField::ShortRepeat => self.short_repeat,
-            ZshOptionField::RcQuotes => self.rc_quotes,
-            ZshOptionField::InteractiveComments => self.interactive_comments,
-            ZshOptionField::CBases => self.c_bases,
-            ZshOptionField::OctalZeroes => self.octal_zeroes,
-        }
-    }
-
-    /// Merge two option snapshots field by field.
-    pub fn merge(&self, other: &Self) -> Self {
-        let mut merged = Self::zsh_default();
-        for field in ZshOptionField::ALL {
-            merged.set_field(field, self.field(field).merge(other.field(field)));
-        }
-        merged
-    }
-
-    fn apply_named_option(&mut self, name: &str, enable: bool) -> bool {
-        let Some((field, value)) = parse_zsh_option_assignment(name, enable) else {
-            return false;
-        };
-        self.set_field(
-            field,
-            if value {
-                OptionValue::On
-            } else {
-                OptionValue::Off
-            },
-        );
-        true
-    }
-}
-
-impl ZshOptionField {
-    const ALL: [Self; 27] = [
-        Self::ShWordSplit,
-        Self::GlobSubst,
-        Self::RcExpandParam,
-        Self::Glob,
-        Self::Nomatch,
-        Self::NullGlob,
-        Self::CshNullGlob,
-        Self::ExtendedGlob,
-        Self::KshGlob,
-        Self::ShGlob,
-        Self::BareGlobQual,
-        Self::GlobDots,
-        Self::Equals,
-        Self::MagicEqualSubst,
-        Self::ShFileExpansion,
-        Self::GlobAssign,
-        Self::IgnoreBraces,
-        Self::IgnoreCloseBraces,
-        Self::BraceCcl,
-        Self::KshArrays,
-        Self::KshZeroSubscript,
-        Self::ShortLoops,
-        Self::ShortRepeat,
-        Self::RcQuotes,
-        Self::InteractiveComments,
-        Self::CBases,
-        Self::OctalZeroes,
-    ];
-}
-
-/// Dialect plus optional zsh option state used to configure the lexer and parser.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ShellProfile {
-    /// Shell dialect to parse.
-    pub dialect: ShellDialect,
-    /// Optional zsh option state, used only for zsh parsing.
-    pub options: Option<ZshOptionState>,
-}
-
-impl ShellProfile {
-    /// Build a native profile for `dialect`.
-    pub fn native(dialect: ShellDialect) -> Self {
-        Self {
-            dialect,
-            options: (dialect == ShellDialect::Zsh).then(ZshOptionState::zsh_default),
-        }
-    }
-
-    /// Build a profile with explicit zsh option state.
-    pub fn with_zsh_options(dialect: ShellDialect, options: ZshOptionState) -> Self {
-        Self {
-            dialect,
-            options: (dialect == ShellDialect::Zsh).then_some(options),
-        }
-    }
-
-    /// Borrow the zsh option state, if this profile carries one.
-    pub fn zsh_options(&self) -> Option<&ZshOptionState> {
-        self.options.as_ref()
-    }
-}
-
-fn parse_zsh_option_assignment(name: &str, enable: bool) -> Option<(ZshOptionField, bool)> {
-    let mut normalized = String::with_capacity(name.len());
-    for ch in name.chars() {
-        if matches!(ch, '_' | '-') {
-            continue;
-        }
-        normalized.push(ch.to_ascii_lowercase());
-    }
-
-    let (normalized, invert) = if let Some(rest) = normalized.strip_prefix("no") {
-        (rest, true)
-    } else {
-        (normalized.as_str(), false)
+#[cfg(test)]
+mod arithmetic_text_helper_tests {
+    use super::{
+        text_is_self_contained_arithmetic_expression,
+        text_looks_like_nontrivial_arithmetic_expression,
     };
 
-    let field = match normalized {
-        "shwordsplit" => ZshOptionField::ShWordSplit,
-        "globsubst" => ZshOptionField::GlobSubst,
-        "rcexpandparam" => ZshOptionField::RcExpandParam,
-        "glob" | "noglob" => ZshOptionField::Glob,
-        "nomatch" => ZshOptionField::Nomatch,
-        "nullglob" => ZshOptionField::NullGlob,
-        "cshnullglob" => ZshOptionField::CshNullGlob,
-        "extendedglob" => ZshOptionField::ExtendedGlob,
-        "kshglob" => ZshOptionField::KshGlob,
-        "shglob" => ZshOptionField::ShGlob,
-        "bareglobqual" => ZshOptionField::BareGlobQual,
-        "globdots" => ZshOptionField::GlobDots,
-        "equals" => ZshOptionField::Equals,
-        "magicequalsubst" => ZshOptionField::MagicEqualSubst,
-        "shfileexpansion" => ZshOptionField::ShFileExpansion,
-        "globassign" => ZshOptionField::GlobAssign,
-        "ignorebraces" => ZshOptionField::IgnoreBraces,
-        "ignoreclosebraces" => ZshOptionField::IgnoreCloseBraces,
-        "braceccl" => ZshOptionField::BraceCcl,
-        "ksharrays" => ZshOptionField::KshArrays,
-        "kshzerosubscript" => ZshOptionField::KshZeroSubscript,
-        "shortloops" => ZshOptionField::ShortLoops,
-        "shortrepeat" => ZshOptionField::ShortRepeat,
-        "rcquotes" => ZshOptionField::RcQuotes,
-        "interactivecomments" => ZshOptionField::InteractiveComments,
-        "cbases" => ZshOptionField::CBases,
-        "octalzeroes" => ZshOptionField::OctalZeroes,
-        _ => return None,
-    };
-
-    Some((field, if invert { !enable } else { enable }))
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct ZshOptionTimeline {
-    initial: ZshOptionState,
-    entries: Arc<[ZshOptionTimelineEntry]>,
-}
-
-#[derive(Debug, Clone)]
-struct ZshOptionTimelineEntry {
-    offset: usize,
-    state: ZshOptionState,
-}
-
-impl ZshOptionTimeline {
-    fn build(input: &str, shell_profile: &ShellProfile) -> Option<Self> {
-        let initial = shell_profile.zsh_options()?.clone();
-        if !might_mutate_zsh_parser_options(input) {
-            return Some(Self {
-                initial,
-                entries: Arc::from([]),
-            });
-        }
-
-        let entries = ZshOptionPrescanner::new(input, initial.clone()).scan();
-        Some(Self {
-            initial,
-            entries: entries.into(),
-        })
+    #[test]
+    fn requires_nontrivial_expressions() {
+        assert!(text_looks_like_nontrivial_arithmetic_expression("1 + 2"));
+        assert!(text_looks_like_nontrivial_arithmetic_expression("arr[1]"));
+        assert!(text_looks_like_nontrivial_arithmetic_expression("++count"));
+        assert!(!text_looks_like_nontrivial_arithmetic_expression("123"));
+        assert!(!text_looks_like_nontrivial_arithmetic_expression("name"));
+        assert!(!text_looks_like_nontrivial_arithmetic_expression(
+            "latest value"
+        ));
     }
 
-    fn options_at(&self, offset: usize) -> &ZshOptionState {
-        let next_index = self.entries.partition_point(|entry| entry.offset <= offset);
-        if next_index == 0 {
-            &self.initial
-        } else {
-            &self.entries[next_index - 1].state
-        }
+    #[test]
+    fn distinguishes_self_contained_expressions() {
+        assert!(text_is_self_contained_arithmetic_expression("1 + 2"));
+        assert!(text_is_self_contained_arithmetic_expression("(1 + 2)"));
+        assert!(!text_is_self_contained_arithmetic_expression("name"));
+        assert!(!text_is_self_contained_arithmetic_expression("arr[1]"));
+        assert!(!text_is_self_contained_arithmetic_expression("foo + 1"));
+        assert!(!text_is_self_contained_arithmetic_expression(
+            "latest value"
+        ));
     }
 }
-
-fn might_mutate_zsh_parser_options(input: &str) -> bool {
-    input.contains("setopt")
-        || input.contains("unsetopt")
-        || input.contains("emulate")
-        || input.contains("set -o")
-        || input.contains("set +o")
-}
-
-#[derive(Debug, Clone)]
-struct ZshOptionPrescanner<'a> {
-    input: &'a str,
-    offset: usize,
-    state: ZshOptionState,
-    entries: Vec<ZshOptionTimelineEntry>,
-}
-
-#[derive(Debug, Clone)]
-enum PrescanToken {
-    Word {
-        text: String,
-        end: usize,
-    },
-    Separator {
-        kind: PrescanSeparator,
-        start: usize,
-        end: usize,
-    },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PrescanSeparator {
-    Newline,
-    Semicolon,
-    Pipe,
-    Ampersand,
-    OpenParen,
-    CloseParen,
-    OpenBrace,
-    CloseBrace,
-}
-
-#[derive(Debug, Clone)]
-struct PrescanLocalScope {
-    saved_state: ZshOptionState,
-    brace_depth: usize,
-    paren_depth: usize,
-    compounds: Vec<PrescanCompound>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PrescanCompound {
-    If,
-    Loop,
-    Case,
-}
-
-impl PrescanLocalScope {
-    fn simple(saved_state: ZshOptionState) -> Self {
-        Self {
-            saved_state,
-            brace_depth: 0,
-            paren_depth: 0,
-            compounds: Vec::new(),
-        }
-    }
-
-    fn brace_group(saved_state: ZshOptionState) -> Self {
-        Self {
-            brace_depth: 1,
-            ..Self::simple(saved_state)
-        }
-    }
-
-    fn subshell(saved_state: ZshOptionState) -> Self {
-        Self {
-            paren_depth: 1,
-            ..Self::simple(saved_state)
-        }
-    }
-
-    fn update_for_command(&mut self, words: &[String]) {
-        let Some(command) = words.first().map(String::as_str) else {
-            return;
-        };
-
-        match command {
-            "if" => self.compounds.push(PrescanCompound::If),
-            "case" => self.compounds.push(PrescanCompound::Case),
-            "for" | "select" | "while" | "until" => {
-                self.compounds.push(PrescanCompound::Loop);
-            }
-            "repeat" if words.iter().any(|word| word == "do") => {
-                self.compounds.push(PrescanCompound::Loop);
-            }
-            "fi" => self.pop_compound(PrescanCompound::If),
-            "done" => self.pop_compound(PrescanCompound::Loop),
-            "esac" => self.pop_compound(PrescanCompound::Case),
-            _ => {}
-        }
-    }
-
-    fn is_complete(&self) -> bool {
-        self.brace_depth == 0 && self.paren_depth == 0 && self.compounds.is_empty()
-    }
-
-    fn pop_compound(&mut self, compound: PrescanCompound) {
-        if self.compounds.last().copied() == Some(compound) {
-            self.compounds.pop();
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PrescanFunctionHeaderState {
-    None,
-    AfterWord,
-    AfterFunctionKeyword,
-    AfterFunctionName,
-    AfterWordOpenParen,
-    AfterFunctionNameOpenParen,
-    ReadyForBrace,
-}
-
-impl<'a> ZshOptionPrescanner<'a> {
-    fn new(input: &'a str, state: ZshOptionState) -> Self {
-        Self {
-            input,
-            offset: 0,
-            state,
-            entries: Vec::new(),
-        }
-    }
-
-    fn scan(mut self) -> Vec<ZshOptionTimelineEntry> {
-        let mut words = Vec::new();
-        let mut command_end = 0usize;
-        let mut local_scopes = Vec::new();
-        let mut function_header = PrescanFunctionHeaderState::None;
-
-        while let Some(token) = self.next_token() {
-            match token {
-                PrescanToken::Word { text, end } => {
-                    if is_prescan_function_body_start(function_header) {
-                        local_scopes.push(PrescanLocalScope::simple(self.state.clone()));
-                        function_header = PrescanFunctionHeaderState::None;
-                    }
-                    command_end = end;
-                    function_header = match function_header {
-                        PrescanFunctionHeaderState::None => {
-                            if text == "function" {
-                                PrescanFunctionHeaderState::AfterFunctionKeyword
-                            } else {
-                                PrescanFunctionHeaderState::AfterWord
-                            }
-                        }
-                        PrescanFunctionHeaderState::AfterFunctionKeyword => {
-                            PrescanFunctionHeaderState::AfterFunctionName
-                        }
-                        _ => PrescanFunctionHeaderState::None,
-                    };
-                    words.push(text);
-                }
-                PrescanToken::Separator { kind, start, end } => {
-                    self.finish_command(&words, command_end.max(start));
-                    if let Some(scope) = local_scopes.last_mut() {
-                        scope.update_for_command(&words);
-                    }
-                    if matches!(
-                        kind,
-                        PrescanSeparator::Newline | PrescanSeparator::Semicolon
-                    ) {
-                        self.restore_completed_local_scopes(&mut local_scopes, end);
-                    }
-                    words.clear();
-                    command_end = end;
-
-                    match kind {
-                        PrescanSeparator::Newline => {
-                            if !matches!(
-                                function_header,
-                                PrescanFunctionHeaderState::AfterFunctionName
-                                    | PrescanFunctionHeaderState::ReadyForBrace
-                            ) {
-                                function_header = PrescanFunctionHeaderState::None;
-                            }
-                        }
-                        PrescanSeparator::Semicolon
-                        | PrescanSeparator::Pipe
-                        | PrescanSeparator::Ampersand => {
-                            function_header = PrescanFunctionHeaderState::None;
-                        }
-                        PrescanSeparator::OpenParen => {
-                            if is_prescan_function_body_start(function_header) {
-                                local_scopes.push(PrescanLocalScope::subshell(self.state.clone()));
-                                function_header = PrescanFunctionHeaderState::None;
-                            } else {
-                                let next_header = match function_header {
-                                    PrescanFunctionHeaderState::AfterWord => {
-                                        PrescanFunctionHeaderState::AfterWordOpenParen
-                                    }
-                                    PrescanFunctionHeaderState::AfterFunctionName => {
-                                        PrescanFunctionHeaderState::AfterFunctionNameOpenParen
-                                    }
-                                    _ => PrescanFunctionHeaderState::None,
-                                };
-                                if !matches!(
-                                    next_header,
-                                    PrescanFunctionHeaderState::AfterWordOpenParen
-                                        | PrescanFunctionHeaderState::AfterFunctionNameOpenParen
-                                ) {
-                                    local_scopes
-                                        .push(PrescanLocalScope::subshell(self.state.clone()));
-                                }
-                                function_header = next_header;
-                            }
-                        }
-                        PrescanSeparator::CloseParen => {
-                            let closes_function_header = matches!(
-                                function_header,
-                                PrescanFunctionHeaderState::AfterWordOpenParen
-                                    | PrescanFunctionHeaderState::AfterFunctionNameOpenParen
-                            );
-                            function_header = if closes_function_header {
-                                PrescanFunctionHeaderState::ReadyForBrace
-                            } else {
-                                PrescanFunctionHeaderState::None
-                            };
-                            if !closes_function_header {
-                                if let Some(scope) = local_scopes.last_mut()
-                                    && scope.paren_depth > 0
-                                {
-                                    scope.paren_depth -= 1;
-                                }
-                                self.restore_completed_local_scopes(&mut local_scopes, end);
-                            }
-                        }
-                        PrescanSeparator::OpenBrace => {
-                            if is_prescan_function_body_start(function_header) {
-                                local_scopes
-                                    .push(PrescanLocalScope::brace_group(self.state.clone()));
-                            } else if let Some(scope) = local_scopes.last_mut() {
-                                scope.brace_depth += 1;
-                            }
-                            function_header = PrescanFunctionHeaderState::None;
-                        }
-                        PrescanSeparator::CloseBrace => {
-                            if let Some(scope) = local_scopes.last_mut()
-                                && scope.brace_depth > 0
-                            {
-                                scope.brace_depth -= 1;
-                            }
-                            self.restore_completed_local_scopes(&mut local_scopes, end);
-                            function_header = PrescanFunctionHeaderState::None;
-                        }
-                    }
-                }
-            }
-        }
-
-        self.finish_command(&words, command_end.max(self.input.len()));
-        if let Some(scope) = local_scopes.last_mut() {
-            scope.update_for_command(&words);
-        }
-        self.restore_completed_local_scopes(&mut local_scopes, self.input.len());
-        self.entries
-    }
-
-    fn finish_command(&mut self, words: &[String], end_offset: usize) {
-        let mut next = self.state.clone();
-        if !apply_prescan_command_effects(words, &mut next) || next == self.state {
-            return;
-        }
-
-        self.state = next.clone();
-        self.entries.push(ZshOptionTimelineEntry {
-            offset: end_offset,
-            state: next,
-        });
-    }
-
-    fn next_token(&mut self) -> Option<PrescanToken> {
-        loop {
-            self.skip_horizontal_whitespace();
-            let ch = self.peek_char()?;
-
-            if ch == '#' && self.state.interactive_comments.is_definitely_on() {
-                self.skip_comment();
-                continue;
-            }
-
-            return match ch {
-                '\n' => {
-                    let start = self.offset;
-                    self.advance_char();
-                    Some(PrescanToken::Separator {
-                        kind: PrescanSeparator::Newline,
-                        start,
-                        end: self.offset,
-                    })
-                }
-                ';' | '|' | '&' | '(' | ')' | '{' | '}' => {
-                    let start = self.offset;
-                    self.advance_char();
-                    if matches!(ch, '|' | '&' | ';') && self.peek_char() == Some(ch) {
-                        self.advance_char();
-                    }
-                    let kind = match ch {
-                        ';' => PrescanSeparator::Semicolon,
-                        '|' => PrescanSeparator::Pipe,
-                        '&' => PrescanSeparator::Ampersand,
-                        '(' => PrescanSeparator::OpenParen,
-                        ')' => PrescanSeparator::CloseParen,
-                        '{' => PrescanSeparator::OpenBrace,
-                        '}' => PrescanSeparator::CloseBrace,
-                        _ => unreachable!(),
-                    };
-                    Some(PrescanToken::Separator {
-                        kind,
-                        start,
-                        end: self.offset,
-                    })
-                }
-                _ => self
-                    .read_word()
-                    .map(|(text, end)| PrescanToken::Word { text, end }),
-            };
-        }
-    }
-
-    fn skip_horizontal_whitespace(&mut self) {
-        while let Some(ch) = self.peek_char() {
-            match ch {
-                ' ' | '\t' => {
-                    self.advance_char();
-                }
-                '\\' if self.second_char() == Some('\n') => {
-                    self.advance_char();
-                    self.advance_char();
-                }
-                _ => break,
-            }
-        }
-    }
-
-    fn skip_comment(&mut self) {
-        while let Some(ch) = self.peek_char() {
-            if ch == '\n' {
-                break;
-            }
-            self.advance_char();
-        }
-    }
-
-    fn read_word(&mut self) -> Option<(String, usize)> {
-        let mut text = String::new();
-
-        while let Some(ch) = self.peek_char() {
-            if is_prescan_separator(ch) {
-                break;
-            }
-
-            match ch {
-                ' ' | '\t' => break,
-                '\\' => {
-                    self.advance_char();
-                    match self.peek_char() {
-                        Some('\n') => {
-                            self.advance_char();
-                        }
-                        Some(next) => {
-                            text.push(next);
-                            self.advance_char();
-                        }
-                        None => text.push('\\'),
-                    }
-                }
-                '\'' => {
-                    self.advance_char();
-                    while let Some(next) = self.peek_char() {
-                        if next == '\'' {
-                            if self.state.rc_quotes.is_definitely_on()
-                                && self.second_char() == Some('\'')
-                            {
-                                text.push('\'');
-                                self.advance_char();
-                                self.advance_char();
-                                continue;
-                            }
-                            self.advance_char();
-                            break;
-                        }
-                        text.push(next);
-                        self.advance_char();
-                    }
-                }
-                '"' => {
-                    self.advance_char();
-                    while let Some(next) = self.peek_char() {
-                        if next == '"' {
-                            self.advance_char();
-                            break;
-                        }
-                        if next == '\\' {
-                            self.advance_char();
-                            if let Some(escaped) = self.peek_char() {
-                                text.push(escaped);
-                                self.advance_char();
-                            }
-                            continue;
-                        }
-                        text.push(next);
-                        self.advance_char();
-                    }
-                }
-                _ => {
-                    text.push(ch);
-                    self.advance_char();
-                }
-            }
-        }
-
-        (!text.is_empty()).then_some((text, self.offset))
-    }
-
-    fn peek_char(&self) -> Option<char> {
-        self.input[self.offset..].chars().next()
-    }
-
-    fn second_char(&self) -> Option<char> {
-        let mut chars = self.input[self.offset..].chars();
-        chars.next()?;
-        chars.next()
-    }
-
-    fn advance_char(&mut self) -> Option<char> {
-        let ch = self.peek_char()?;
-        self.offset += ch.len_utf8();
-        Some(ch)
-    }
-
-    fn restore_completed_local_scopes(
-        &mut self,
-        local_scopes: &mut Vec<PrescanLocalScope>,
-        offset: usize,
-    ) {
-        while local_scopes
-            .last()
-            .is_some_and(PrescanLocalScope::is_complete)
-        {
-            let Some(scope) = local_scopes.pop() else {
-                unreachable!("scope just matched");
-            };
-            if self.state != scope.saved_state {
-                self.state = scope.saved_state.clone();
-                self.entries.push(ZshOptionTimelineEntry {
-                    offset,
-                    state: scope.saved_state,
-                });
-            } else {
-                self.state = scope.saved_state;
-            }
-        }
-    }
-}
-
-fn is_prescan_separator(ch: char) -> bool {
-    matches!(ch, '\n' | ';' | '|' | '&' | '(' | ')' | '{' | '}')
-}
-
-fn is_prescan_function_body_start(state: PrescanFunctionHeaderState) -> bool {
-    matches!(
-        state,
-        PrescanFunctionHeaderState::AfterFunctionName | PrescanFunctionHeaderState::ReadyForBrace
-    )
-}
-
-fn apply_prescan_command_effects(words: &[String], state: &mut ZshOptionState) -> bool {
-    let Some((command, args_index)) = normalize_prescan_command(words) else {
-        return false;
-    };
-
-    match command {
-        "setopt" => {
-            let mut changed = false;
-            for arg in &words[args_index..] {
-                changed |= state.apply_setopt(arg);
-            }
-            changed
-        }
-        "unsetopt" => {
-            let mut changed = false;
-            for arg in &words[args_index..] {
-                changed |= state.apply_unsetopt(arg);
-            }
-            changed
-        }
-        "set" => apply_prescan_set_builtin(&words[args_index..], state),
-        "emulate" => apply_prescan_emulate(&words[args_index..], state),
-        _ => false,
-    }
-}
-
-fn normalize_prescan_command(words: &[String]) -> Option<(&str, usize)> {
-    let mut index = 0usize;
-
-    while let Some(word) = words.get(index) {
-        if is_prescan_assignment_word(word) {
-            index += 1;
-            continue;
-        }
-        match static_command_wrapper_target_index(words.len(), index, word, |word_index| {
-            Some(Cow::Borrowed(words[word_index].as_str()))
-        }) {
-            StaticCommandWrapperTarget::NotWrapper => {}
-            StaticCommandWrapperTarget::Wrapper {
-                target_index: Some(target_index),
-            } => {
-                index = target_index;
-                continue;
-            }
-            StaticCommandWrapperTarget::Wrapper { target_index: None } => return None,
-        }
-        return Some((word.as_str(), index + 1));
-    }
-
-    None
-}
-
-fn is_prescan_assignment_word(word: &str) -> bool {
-    let Some((name, _value)) = word.split_once('=') else {
-        return false;
-    };
-    !name.is_empty()
-        && !name.starts_with('-')
-        && name
-            .chars()
-            .all(|ch| ch.is_ascii_alphanumeric() || ch == '_')
-}
-
-fn apply_prescan_set_builtin(words: &[String], state: &mut ZshOptionState) -> bool {
-    let mut changed = false;
-    let mut index = 0usize;
-
-    while let Some(word) = words.get(index) {
-        match word.as_str() {
-            "-o" | "+o" => {
-                let enable = word.starts_with('-');
-                if let Some(name) = words.get(index + 1) {
-                    changed = if enable {
-                        state.apply_setopt(name)
-                    } else {
-                        state.apply_unsetopt(name)
-                    } || changed;
-                }
-                index += 2;
-            }
-            _ => {
-                if let Some(name) = word.strip_prefix("-o") {
-                    changed = state.apply_setopt(name) || changed;
-                } else if let Some(name) = word.strip_prefix("+o") {
-                    changed = state.apply_unsetopt(name) || changed;
-                }
-                index += 1;
-            }
-        }
-    }
-
-    changed
-}
-
-fn apply_prescan_emulate(words: &[String], state: &mut ZshOptionState) -> bool {
-    let mut changed = false;
-    let mut mode = None;
-    let mut pending_option: Option<bool> = None;
-    let mut explicit_updates = Vec::new();
-    let mut index = 0usize;
-
-    while let Some(word) = words.get(index) {
-        if let Some(enable) = pending_option.take() {
-            explicit_updates.push((word.clone(), enable));
-            index += 1;
-            continue;
-        }
-
-        match word.as_str() {
-            "-o" | "+o" => {
-                pending_option = Some(word.starts_with('-'));
-                index += 1;
-                continue;
-            }
-            "zsh" | "sh" | "ksh" | "csh" if mode.is_none() => {
-                mode = Some(match word.as_str() {
-                    "zsh" => ZshEmulationMode::Zsh,
-                    "sh" => ZshEmulationMode::Sh,
-                    "ksh" => ZshEmulationMode::Ksh,
-                    "csh" => ZshEmulationMode::Csh,
-                    _ => unreachable!(),
-                });
-                index += 1;
-                continue;
-            }
-            _ if mode.is_none() && word.starts_with('-') => {
-                for flag in word[1..].chars() {
-                    if flag == 'o' {
-                        pending_option = Some(true);
-                    }
-                }
-                index += 1;
-                continue;
-            }
-            _ => {}
-        }
-
-        index += 1;
-    }
-
-    if let Some(mode) = mode {
-        *state = ZshOptionState::for_emulate(mode);
-        changed = true;
-    }
-
-    for (name, enable) in explicit_updates {
-        changed = if enable {
-            state.apply_setopt(&name)
-        } else {
-            state.apply_unsetopt(&name)
-        } || changed;
-    }
-
-    changed
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct DialectFeatures {
-    double_bracket: bool,
-    arithmetic_command: bool,
-    arithmetic_for: bool,
-    function_keyword: bool,
-    select_loop: bool,
-    coproc_keyword: bool,
-    zsh_repeat_loop: bool,
-    zsh_foreach_loop: bool,
-    zsh_parameter_modifiers: bool,
-    zsh_brace_if: bool,
-    zsh_always: bool,
-    zsh_background_operators: bool,
-    zsh_glob_qualifiers: bool,
-}
-
-impl ShellDialect {
-    /// Infer a shell dialect from a command or shebang name.
-    pub fn from_name(name: &str) -> Self {
-        match name.trim().to_ascii_lowercase().as_str() {
-            "sh" | "dash" | "ksh" | "posix" => Self::Posix,
-            "mksh" => Self::Mksh,
-            "zsh" => Self::Zsh,
-            _ => Self::Bash,
-        }
-    }
-
-    const fn features(self) -> DialectFeatures {
-        match self {
-            Self::Posix => DialectFeatures {
-                double_bracket: false,
-                arithmetic_command: false,
-                arithmetic_for: false,
-                function_keyword: true,
-                select_loop: false,
-                coproc_keyword: false,
-                zsh_repeat_loop: false,
-                zsh_foreach_loop: false,
-                zsh_parameter_modifiers: false,
-                zsh_brace_if: false,
-                zsh_always: false,
-                zsh_background_operators: false,
-                zsh_glob_qualifiers: false,
-            },
-            Self::Mksh => DialectFeatures {
-                double_bracket: true,
-                arithmetic_command: true,
-                arithmetic_for: false,
-                function_keyword: true,
-                select_loop: true,
-                coproc_keyword: false,
-                zsh_repeat_loop: false,
-                zsh_foreach_loop: false,
-                zsh_parameter_modifiers: false,
-                zsh_brace_if: false,
-                zsh_always: false,
-                zsh_background_operators: false,
-                zsh_glob_qualifiers: false,
-            },
-            Self::Bash => DialectFeatures {
-                double_bracket: true,
-                arithmetic_command: true,
-                arithmetic_for: true,
-                function_keyword: true,
-                select_loop: true,
-                coproc_keyword: true,
-                zsh_repeat_loop: false,
-                zsh_foreach_loop: false,
-                zsh_parameter_modifiers: false,
-                zsh_brace_if: false,
-                zsh_always: false,
-                zsh_background_operators: false,
-                zsh_glob_qualifiers: false,
-            },
-            Self::Zsh => DialectFeatures {
-                double_bracket: true,
-                arithmetic_command: true,
-                arithmetic_for: true,
-                function_keyword: true,
-                select_loop: true,
-                coproc_keyword: true,
-                zsh_repeat_loop: true,
-                zsh_foreach_loop: true,
-                zsh_parameter_modifiers: true,
-                zsh_brace_if: true,
-                zsh_always: true,
-                zsh_background_operators: true,
-                zsh_glob_qualifiers: true,
-            },
-        }
-    }
-}
-
-/// Parser for bash scripts.
-#[derive(Clone)]
-pub struct Parser<'a> {
-    input: &'a str,
-    lexer: Lexer<'a>,
-    synthetic_tokens: VecDeque<SyntheticToken>,
-    alias_replays: Vec<AliasReplay>,
-    current_token: Option<LexedToken<'a>>,
-    current_word_cache: Option<Word>,
-    current_token_kind: Option<TokenKind>,
-    current_keyword: Option<Keyword>,
-    /// Span of the current token
-    current_span: Span,
-    /// Lookahead token for function parsing
-    peeked_token: Option<LexedToken<'a>>,
-    /// Maximum allowed AST nesting depth
-    max_depth: usize,
-    /// Current nesting depth
-    current_depth: usize,
-    /// Remaining fuel for parsing operations
-    fuel: usize,
-    /// Maximum fuel (for error reporting)
-    max_fuel: usize,
-    /// Comments collected during parsing.
-    comments: Vec<Comment>,
-    /// Known aliases declared earlier in the current parse stream.
-    aliases: HashMap<String, AliasDefinition>,
-    /// Whether alias expansion is currently enabled.
-    expand_aliases: bool,
-    /// Whether the next fetched word is eligible for alias expansion because
-    /// the previous alias expansion ended with trailing whitespace.
-    expand_next_word: bool,
-    /// Nesting depth of active brace-delimited statement sequences.
-    brace_group_depth: usize,
-    /// Active brace-body parsing contexts, used to distinguish compact zsh
-    /// closers from literal `}` arguments.
-    brace_body_stack: Vec<BraceBodyContext>,
-    syntax_facts: SyntaxFacts,
-    shell_profile: ShellProfile,
-    zsh_timeline: Option<Arc<ZshOptionTimeline>>,
-    dialect: ShellDialect,
-    #[cfg(feature = "benchmarking")]
-    benchmark_counters: Option<ParserBenchmarkCounters>,
-}
-
-#[derive(Clone)]
-struct ParserCheckpoint<'a> {
-    lexer: Lexer<'a>,
-    synthetic_tokens: VecDeque<SyntheticToken>,
-    alias_replays: Vec<AliasReplay>,
-    current_token: Option<LexedToken<'a>>,
-    current_token_kind: Option<TokenKind>,
-    current_keyword: Option<Keyword>,
-    current_span: Span,
-    peeked_token: Option<LexedToken<'a>>,
-    current_depth: usize,
-    fuel: usize,
-    comments: Vec<Comment>,
-    expand_next_word: bool,
-    brace_group_depth: usize,
-    brace_body_stack: Vec<BraceBodyContext>,
-    syntax_facts: SyntaxFacts,
-    #[cfg(feature = "benchmarking")]
-    benchmark_counters: Option<ParserBenchmarkCounters>,
-}
-
-/// A parser diagnostic emitted while recovering from invalid input.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ParseDiagnostic {
-    /// Human-readable diagnostic message.
-    pub message: String,
-    /// Source span associated with the diagnostic.
-    pub span: Span,
-}
-
-#[derive(Debug, Clone)]
-struct AliasDefinition {
-    tokens: Arc<[LexedToken<'static>]>,
-    expands_next_word: bool,
-}
-
-#[derive(Debug, Clone)]
-struct AliasReplay {
-    tokens: Arc<[LexedToken<'static>]>,
-    next_index: usize,
-    base: Position,
-}
-
-impl AliasReplay {
-    fn new(alias: &AliasDefinition, base: Position) -> Self {
-        Self {
-            tokens: Arc::clone(&alias.tokens),
-            next_index: 0,
-            base,
-        }
-    }
-
-    fn next_token<'b>(&mut self) -> Option<LexedToken<'b>> {
-        let token = self.tokens.get(self.next_index)?.clone();
-        self.next_index += 1;
-        Some(token.into_owned().rebased(self.base).with_synthetic_flag())
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct SyntheticToken {
-    kind: TokenKind,
-    span: Span,
-}
-
-impl SyntheticToken {
-    const fn punctuation(kind: TokenKind, span: Span) -> Self {
-        Self { kind, span }
-    }
-
-    fn materialize<'b>(self) -> LexedToken<'b> {
-        LexedToken::punctuation(self.kind).with_span(self.span)
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-enum FlowControlBuiltinKind {
-    Break,
-    Continue,
-    Return,
-    Exit,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BraceBodyContext {
-    Ordinary,
-    Function,
-    IfClause,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct TokenSet(u64);
-
-impl TokenSet {
-    const fn contains(self, kind: TokenKind) -> bool {
-        self.0 & (1u64 << kind as u8) != 0
-    }
-}
-
-macro_rules! token_set {
-    ($($kind:path),+ $(,)?) => {
-        TokenSet(0 $(| (1u64 << ($kind as u8)))+)
-    };
-}
-
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Keyword {
-    If,
-    For,
-    Repeat,
-    Foreach,
-    While,
-    Until,
-    Case,
-    Select,
-    Time,
-    Coproc,
-    Function,
-    Always,
-    Then,
-    Else,
-    Elif,
-    Fi,
-    Do,
-    Done,
-    Esac,
-    In,
-}
-
-impl Keyword {
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::If => "if",
-            Self::For => "for",
-            Self::Repeat => "repeat",
-            Self::Foreach => "foreach",
-            Self::While => "while",
-            Self::Until => "until",
-            Self::Case => "case",
-            Self::Select => "select",
-            Self::Time => "time",
-            Self::Coproc => "coproc",
-            Self::Function => "function",
-            Self::Always => "always",
-            Self::Then => "then",
-            Self::Else => "else",
-            Self::Elif => "elif",
-            Self::Fi => "fi",
-            Self::Do => "do",
-            Self::Done => "done",
-            Self::Esac => "esac",
-            Self::In => "in",
-        }
-    }
-}
-
-impl std::fmt::Display for Keyword {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct KeywordSet(u32);
-
-impl KeywordSet {
-    const fn single(keyword: Keyword) -> Self {
-        Self(1u32 << keyword as u8)
-    }
-
-    const fn contains(self, keyword: Keyword) -> bool {
-        self.0 & (1u32 << keyword as u8) != 0
-    }
-}
-
-macro_rules! keyword_set {
-    ($($keyword:ident),+ $(,)?) => {
-        KeywordSet(0 $(| (1u32 << (Keyword::$keyword as u8)))+)
-    };
-}
-
-const PIPE_OPERATOR_TOKENS: TokenSet = token_set![TokenKind::Pipe, TokenKind::PipeBoth];
-const REDIRECT_TOKENS: TokenSet = token_set![
-    TokenKind::RedirectOut,
-    TokenKind::Clobber,
-    TokenKind::RedirectAppend,
-    TokenKind::RedirectIn,
-    TokenKind::RedirectReadWrite,
-    TokenKind::HereString,
-    TokenKind::HereDoc,
-    TokenKind::HereDocStrip,
-    TokenKind::RedirectBoth,
-    TokenKind::RedirectBothAppend,
-    TokenKind::DupOutput,
-    TokenKind::RedirectFd,
-    TokenKind::RedirectFdAppend,
-    TokenKind::DupFd,
-    TokenKind::DupInput,
-    TokenKind::DupFdIn,
-    TokenKind::DupFdClose,
-    TokenKind::RedirectFdIn,
-    TokenKind::RedirectFdReadWrite,
-];
-const NON_COMMAND_KEYWORDS: KeywordSet =
-    keyword_set![Then, Else, Elif, Fi, Do, Done, Esac, In, Always];
-const IF_BODY_TERMINATORS: KeywordSet = keyword_set![Elif, Else, Fi];
 
 impl<'a> Parser<'a> {
     /// Create a new parser for the given input.
